+++ src/ModularPipelines.Build/Modules/UpdateReleaseNotesModule.cs
@@ -75,7 +75,7 @@ public class UpdateReleaseNotesModule : Module
         await releaseNotesFile.WriteAsync(string.Empty);
         
         await GitHelpers.SetUserCommitInformation(context, cancellationToken);
-        await GitHelpers.SetTarget(context, "main", _githubSettings.Value.TokenWithoutTriggerBuild!, cancellationToken);
+        await GitHelpers.SetToken(context, _githubSettings.Value.TokenWithoutTriggerBuild!, cancellationToken);
         await GitHelpers.CommitAndPush(context, "main", "Create Release", cancellationToken);
     }
 }