+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,3 +1,1 @@
-- Added Release Notes Generator
-- Added ConsoleUse Analyzer / Detector
-- Reworked Environment Variables in the Build pipeline
+- Fix setting token on Git push